+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the desktop search form to the `Header` component with the `SearchButton` and `SearchForm` components.
 - Refactors `SearchButton` and `SearchForm` to implement the mobile search form.
 - Adds the mobile navigation menu for the `Header` component with `MobileNav` and `MobileNavButton` components.
+- Adds the mobile login menu for the `Header` component.
 
 ## Prerelease